Move more classes into separate files :truck:
