+++ Src/Notion.Client/Api/Blocks/RequestParams/BlocksRetrieveChildrenParameters.cs
@@ -1,9 +1,5 @@
 namespace Notion.Client
 {
-    public interface IBlocksRetrieveChildrenQueryParameters : IPaginationParameters
-    {
-    }
-
     public class BlocksRetrieveChildrenParameters : IBlocksRetrieveChildrenQueryParameters
     {
         public string StartCursor { get; set; }
