+++ Src/Notion.Client/Api/Blocks/RequestParams/IBlocksAppendChildrenBodyParameters.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Notion.Client
+{
+    // TODO: need an input version of Block
+    public interface IBlocksAppendChildrenBodyParameters
+    {
+        IEnumerable<Block> Children { get; set; }
+    }
+}
