+++ Src/Notion.Client/Api/Blocks/RequestParams/IBlocksRetrieveChildrenQueryParameters.cs
@@ -1,0 +1,6 @@
+﻿namespace Notion.Client
+{
+    public interface IBlocksRetrieveChildrenQueryParameters : IPaginationParameters
+    {
+    }
+}
