+++ Src/Notion.Client/Api/Databases/RequestParams/Direction.cs
@@ -1,0 +1,16 @@
+﻿using System.Runtime.Serialization;
+
+namespace Notion.Client
+{
+    public enum Direction
+    {
+        [EnumMember(Value = null)]
+        Unknown,
+
+        [EnumMember(Value = "ascending")]
+        Ascending,
+
+        [EnumMember(Value = "descending")]
+        Descending
+    }
+}
