+++ Src/Notion.Client/Api/Users/IUsersClient.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+
+namespace Notion.Client
+{
+    public interface IUsersClient
+    {
+        Task<User> RetrieveAsync(string userId);
+        Task<PaginatedList<User>> ListAsync();
+    }
+}
